Added files via upload
This grabs the active drivers and active clients from database, creates array with driver username, latitude & longitude then isolates all client latitude and longitude with client username in tow. It calls Google Direction Matrix with Driver as origin and Client as destination which returns the calculated JSON. It parses the JSON isolating the time. Then it stores it into an array with client user name. Sorts the array by time ascending and then picks off the top array element. 

I haven't made it execute or update the database using Vivi's procedure. It just tells which driver to be paired with passenger. The find is slow because it does multiple queries to Google Maps, so I think one way to resolve is to make it just do 1 query. I know it's possible but will be a little bit of a headache keeping organized who I just searched for with whom. Whatever.